third_party/blink/web_tests/virtual/html-anchor-attribute-disabled/external/wpt/html/semantics/popovers/popover-anchor-multicol-display.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] .target 1
  assert_equals: \n<div class="target" popover="" anchor="anchor1" data-expected-height="30"></div>\nheight expected 30 but got 20
Harness: the test ran to completion.
